requirements.txt
asgiref==3.4.1
Django==3.2.5
django-leaflet==0.28.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
